requirements.txt
cvxpy==1.4.1
matplotlib==3.8.2
commonroad-all==0.0.1
commonroad-drivability-checker==2022.2.1
commonroad-io==2023.1
commonroad-route-planner==2022.3
commonroad-vehicle-models==3.0.2
